+++ Percolation/src/main/java/com/pavelhudau/percolation/Main.java
@@ -1,8 +1,8 @@
-package main.java;
+package com.pavelhudau.percolation;
 
 public class Main {
 
     public static void main(String[] args) {
-	// write your code here
+        // write your code here
     }
 }
