+++ Percolation/src/main/java/com/pavelhudau/percolation/Percolation.java
@@ -1,4 +1,4 @@
-package main.java;
+package com.pavelhudau.percolation;
 
 import java.util.HashMap;
 import java.util.Map;
