+++ Percolation/src/main/java/com/pavelhudau/percolation/PercolationStats.java
@@ -1,4 +1,4 @@
-package main.java;
+package com.pavelhudau.percolation;
 
 import javafx.util.Pair;
