+++ Percolation/src/test/java/com/pavelhudau/percolation/PercolationTest.java
@@ -1,6 +1,5 @@
-package test.java;
+package com.pavelhudau.percolation;
 
-import main.java.Percolation;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
